fix warnings
- add e_game_playback_type

+++ xlive/Blam/Cache/DataTypes/TagBlock.h
@@ -3,8 +3,8 @@
 template<typename T = void>
 struct tag_block
 {
-	size_t size;
-	size_t data;
+	int32 size;
+	int32 data;
 
 	int data_size() const
 	{
@@ -34,7 +34,7 @@ struct tag_block
 			return nullptr;
 	}
 
-	T* operator[](size_t index) const
+	T* operator[](int32 index) const
 	{
 		static_assert(std::is_void<T>::value == false, "You need to set the tag block type to use this function");
 		if (index == NONE)
